backend/account/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
databases[mysql]
aiomysql
passlib
pyjwt